Lint Phase 2: fix bugprone-narrowing-conversions in the ambiguity-frequency chain at its source
nucleotide2ambiguity_nucleotide()/DNA_b128_String::calcAmbiguityProbabilities()
took int basefreqA/C/G/T while every real caller passes DNA_b128_String's
size_t num_As_/num_Cs_/num_Gs_/num_Ts_ counts - verified the function
bodies only ever add/multiply/divide these (never subtract, so no
signed-vs-unsigned wraparound risk), then widened the parameters to
size_t and introduced local `const double A/C/G/T` aliases for the
probability-ratio arithmetic (that arithmetic already needs double
precision, and size_t->double is itself a narrowing case bugprone
flags for very large values - irrelevant here but resolved with an
explicit cast anyway, at the one place the values are used instead of
at each of the ~130 case-label expressions).

This is the "fix the type at the declaration, it's a small local
change" side of Phase 2's narrowing-conversions triage. The dominant
remaining cluster (DistanceMatrix::getDistance/setDistance's int i,j
receiving size_t indices throughout the codebase, and
compute_Tamura_Nei's int strlen/numAs/etc, which - unlike this chain -
does subtract (strlen - sd.deletedPositions), so its parameter types
should NOT be widened) needs per-call-site explicit casts instead,
which is Phase 2's next, much larger chunk of work.

Verified via full rebuild + ctest + RunExamples.sh, all 15 fixtures
byte-identical to expected_output/.

+++ include/fastphylo/dna/DNA_b128_String.hpp
@@ -195,8 +195,8 @@ public:
   //These functions compute the probabilities of every ambiguity symbol.
   //E.g. 'M' is either an 'A' or a 'C' and if basefreqA=3 and basefreqC=2
   //then the probability of seeing an 'A' is 3/5 and for 'C' 2/5.
-  void calcAmbiguityProbabilities(int basefreqA, int basefreqC,
-                                  int basefreqG, int basefreqT){
+  void calcAmbiguityProbabilities(size_t basefreqA, size_t basefreqC,
+                                  size_t basefreqG, size_t basefreqT){
     std::vector<ambiguity_nucleotide_at_position>::iterator iter = ambiguities.begin();
     for ( ; iter != ambiguities.end() ; ++iter ){
       (*iter).ambiguity = nucleotide2ambiguity_nucleotide((*iter).ambiguity.n, basefreqA, basefreqC, basefreqG, basefreqT);
